+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - **Import Command**
   - `spout import` now changes variables with all caps `display_name` to use title case instead
     - Display names that use mixed case are unaffected
+- **Testing Changes**
+  - Tests now include checks to assure that variable display_name fields don't exceed 255 length requirement
+    - `include Spout::Tests::VariableDisplayNameLength`
 
 ## 0.9.1 (October 14, 2014)